ENH: Implement global delete as a specialization of an empty change
Add a:isDelete flag to ChangeGlobally#SetParameters().
Evaluate the flag and skip the arming of insert mode to implement delete as a change with an empty replacement.
For linewise delete, don't remove the trailing newline in the search pattern, so that the entire line(s) are removed, and no blank line is left.