FIX: Minor: Last keyword in a line is not found when the cursor already is inside
I think this has to do with the fact that search() always starts from the start, but discards matches before the cursor position. Now \k\+ makes up an entire match, and no retry happens.
We could solve this by supplying the 'z' flag, but as that's newish, do a backward-compatible fix and search for an individual keyword character, as we only need to establish the start position of a match here, anyway (the grabbing is done by the a:sourceTextObject).