Enable hooks to influence the ChangeGlobally#AreaReplaceSecondPass()
As this is where the real replacement happens for the area mappings, and which ChangeGloballySmartCase.vim needs to manipulate. By passing the second pass call into the first pass (but ignoring it), and having the logic in ChangeGlobally#CountedAreaReplace() extract it instead of hard-coding, we enable the hook to reuse most of the existing functionality, and don't have to open additional seams.