Support {target-motion} of gv in <Leader>gc and <Leader>gx to replace in the last selection
Previously, the {source-motion} clobbered the user-established visual selection. Now, instead of defining a visual selection to hold the source motion, store it in a script variable, invoke a new ChangeGlobally#AreaSourceOperatorTarget() as the opfunc for the {target-motion}, and then define the visual selection there (i.e. _after_ the second opfunc was executed instead of _before_). We can then still delegate the rest to ChangeGlobally#SelectionSourceOperatorTarget(), which has target in '[,'] and source in '<,'>.