Refactoring: Yank instead of deleting changed text for gx mappings
There's no need to delete in order to obtain the changed text, undo that delete, and then do the (counted) delete again... just yank the text. It's just that the initial implementation was for gc mappings, and gx was added afterwards.
Rename s:DeleteChangedText() to s:GetChangedText() and pass "y" instead of "d" if s:isDelete is true.
Move the undo handling inside the ! s:isDelete conditional in ChangeGlobally#Substitute().